Add files via upload
A Hospital Management System (HMS) with Chatbot is a comprehensive digital solution designed to streamline the administrative, clinical, and support processes in a hospital or clinic, while enhancing user interaction through an AI-powered chatbot. This system integrates traditional hospital management functionalities with intelligent conversational interfaces to improve the efficiency and accessibility of healthcare services.